+++ packages/web-vue/lib/slider/index.vue
@@ -11,9 +11,8 @@
 		ref="sliderRef"
 		@click.self="selectSliderHandler"
 		@mousedown="focusSliderHandler"
-		:tabindex="disabledComputed || readonlyComputed ? -1 : 0"
-		@focusin="focusinHandler"
-		@focusout="focusoutHandler"
+		@focusin="focusHandler"
+		@focusout="blurHandler"
 	>
 		<div class="px-slider-mark-wrapper" ref="markWrapperRef">
 			<template v-if="markPoints.length">
@@ -43,7 +42,8 @@
 			:content-props="{
 				onMousedown: numberDisplayMousedownHandler
 			}"
-			v-if="props.tooltip && !props.range"
+			:visible="props.tooltip ? undefined : false"
+			v-if="!props.range"
 		>
 			<div
 				class="px-slider-thumb"
@@ -53,6 +53,8 @@
 				}"
 				@mousedown="startDrag"
 				@touchstart.passive="startDrag"
+				@focusin="thumbFocusHandler"
+				@focusout="thumbBlurHandler"
 				:tabindex="disabledComputed || readonlyComputed ? -1 : 0"
 				ref="thumbRef"
 			>
@@ -67,22 +69,7 @@
 				</div>
 			</template>
 		</Popup>
-		<div
-			v-if="!props.range && !props.tooltip"
-			class="px-slider-thumb"
-			:style="{
-				left: props.direction === 'horizontal' ? thumbLeft + 'px' : undefined,
-				bottom: props.direction === 'horizontal' ? undefined : thumbLeft + 'px'
-			}"
-			@mousedown="startDrag"
-			@touchstart.passive="startDrag"
-			:tabindex="disabledComputed || readonlyComputed ? -1 : 0"
-			ref="thumbRef"
-		>
-			<slot name="thumb"></slot>
-			<canvas ref="thumbCanvasRef" class="px-slider-canvas"></canvas>
-		</div>
-		<template v-if="props.range">
+		<template v-else>
 			<Popup
 				:arrow="false"
 				:placement="props.direction === 'horizontal' ? 'top' : 'right'"
@@ -91,7 +78,7 @@
 				:content-props="{
 					onMousedown: numberDisplayMousedownHandler
 				}"
-				v-if="props.tooltip"
+				:visible="props.tooltip ? undefined : false"
 			>
 				<div
 					class="px-slider-thumb px-slider-thumb-start"
@@ -102,6 +89,8 @@
 					@mousedown="(e) => startDrag(e, 'start')"
 					@touchstart.passive="(e) => startDrag(e, 'start')"
 					:tabindex="disabledComputed || readonlyComputed ? -1 : 0"
+					@focusin="thumbStartFocusHandler"
+					@focusout="thumbStartBlurHandler"
 					ref="thumbStartRef"
 				>
 					<slot name="thumb-start"></slot>
@@ -115,22 +104,6 @@
 					</div>
 				</template>
 			</Popup>
-			<div
-				v-else
-				class="px-slider-thumb px-slider-thumb-start"
-				:style="{
-					left: props.direction === 'horizontal' ? thumbLeftStart + 'px' : undefined,
-					bottom: props.direction === 'horizontal' ? undefined : thumbLeftStart + 'px'
-				}"
-				@mousedown="(e) => startDrag(e, 'start')"
-				@touchstart.passive="(e) => startDrag(e, 'start')"
-				:tabindex="disabledComputed || readonlyComputed ? -1 : 0"
-				ref="thumbStartRef"
-			>
-				<slot name="thumb-start"></slot>
-				<canvas ref="thumbStartCanvasRef" class="px-slider-canvas"></canvas>
-			</div>
-
 			<Popup
 				:arrow="false"
 				:placement="props.direction === 'horizontal' ? 'top' : 'right'"
@@ -139,7 +112,7 @@
 				:content-props="{
 					onMousedown: numberDisplayMousedownHandler
 				}"
-				v-if="props.tooltip"
+				:visible="props.tooltip ? undefined : false"
 			>
 				<div
 					class="px-slider-thumb px-slider-thumb-end"
@@ -150,6 +123,8 @@
 					@mousedown="(e) => startDrag(e, 'end')"
 					@touchstart.passive="(e) => startDrag(e, 'end')"
 					:tabindex="disabledComputed || readonlyComputed ? -1 : 0"
+					@focusin="thumbEndFocusHandler"
+					@focusout="thumbEndBlurHandler"
 					ref="thumbEndRef"
 				>
 					<slot name="thumb-end"></slot>
@@ -163,21 +138,6 @@
 					</div>
 				</template>
 			</Popup>
-			<div
-				v-else
-				class="px-slider-thumb px-slider-thumb-end"
-				:style="{
-					left: props.direction === 'horizontal' ? thumbLeftEnd + 'px' : undefined,
-					bottom: props.direction === 'horizontal' ? undefined : thumbLeftEnd + 'px'
-				}"
-				@mousedown="(e) => startDrag(e, 'end')"
-				@touchstart.passive="(e) => startDrag(e, 'end')"
-				:tabindex="disabledComputed || readonlyComputed ? -1 : 0"
-				ref="thumbEndRef"
-			>
-				<slot name="thumb-end"></slot>
-				<canvas ref="thumbEndCanvasRef" class="px-slider-canvas"></canvas>
-			</div>
 		</template>
 		<canvas ref="canvasRef" class="px-slider-canvas"></canvas>
 		<canvas ref="dotCanvasRef" class="px-slider-canvas"></canvas>
@@ -188,9 +148,9 @@
 import {
 	clone,
 	isArray,
-	isNullish,
 	isNumber,
 	isUndefined,
+	range,
 	throttle,
 	type Nullish
 } from 'parsnip-kit'
@@ -221,7 +181,7 @@ import {
 import Popup from '../popup/index.vue'
 import { useTransitionEnd } from '../share/hook/use-transition-end'
 import { usePolling } from '../share/hook/use-polling'
-import { useCancelableDelay } from '../share/hook/use-cancelable-delay'
+import { useFocusMode } from '../share/hook/use-focus-mode'
 
 defineOptions({
 	name: 'Slider'
@@ -260,7 +220,6 @@ const [modelValue, updateModelValue] = useControlledMode('modelValue', props, em
 const thumbRef = shallowRef<null | HTMLDivElement>(null)
 const thumbStartRef = shallowRef<null | HTMLDivElement>(null)
 const thumbEndRef = shallowRef<null | HTMLDivElement>(null)
-const markWrapperRef = shallowRef<null | HTMLDivElement>(null)
 
 const thumbCanvasRef = shallowRef<null | HTMLCanvasElement>(null)
 const thumbStartCanvasRef = shallowRef<null | HTMLCanvasElement>(null)
@@ -563,56 +522,47 @@ const focusSliderHandler = (e: MouseEvent) => {
 	}
 	const target = e.target as HTMLElement
 	setTimeout(() => {
-		if (!isNullish(target.tabIndex)) {
+		if (target.tabIndex >= 0) {
 			target.focus()
-		} else if (target === sliderRef.value && sliderRef.value) {
-			sliderRef.value.focus()
-		} else if (
-			markWrapperRef.value &&
-			markWrapperRef.value.contains(target) &&
-			target !== markWrapperRef.value
-		) {
-			;[...markWrapperRef.value.children]
-				.filter((e) => e instanceof HTMLDivElement && !isNullish(e.tabIndex))
-				.some((e) => {
-					if (e.contains(target)) {
-						;(e as HTMLDivElement).focus()
-						return true
-					}
-				})
 		} else if (thumbRef.value && thumbRef.value.contains(target)) {
 			thumbRef.value.focus()
 		} else if (thumbStartRef.value && thumbStartRef.value.contains(target)) {
 			thumbStartRef.value.focus()
 		} else if (thumbEndRef.value && thumbEndRef.value.contains(target)) {
 			thumbEndRef.value.focus()
-		} else if (sliderRef.value) {
-			sliderRef.value.focus()
+		} else {
+			;(thumbStartRef.value || thumbRef.value)?.focus()
 		}
 	}, 0)
 }
 
-const [wait, cancel] = useCancelableDelay()
-
-const focusState = ref(false)
-
-const focusinHandler = (e: FocusEvent) => {
-	cancel()
-	const currentFocus = focusState.value
-	focusState.value = true
-	if (!currentFocus) {
-		emits('focus', e)
+const { focusHandler, blurHandler, cancel } = useFocusMode({
+	onFocus: (e, isFirstFocus) => {
+		if (isFirstFocus) {
+			emits('focus', e)
+		}
+	},
+	onBlur: (e) => {
+		emits('blur', e)
+		formItemProvide?.blurHandler()
 	}
-}
-const focusoutHandler = async (e: FocusEvent) => {
-	const next = await wait()
-	if (!next) {
-		return
-	}
-	focusState.value = false
-	emits('blur', e)
-	formItemProvide?.blurHandler()
-}
+})
+
+const {
+	focusMode: thumbFocusMode,
+	focusHandler: thumbFocusHandler,
+	blurHandler: thumbBlurHandler
+} = useFocusMode()
+const {
+	focusMode: thumbStartFocusMode,
+	focusHandler: thumbStartFocusHandler,
+	blurHandler: thumbStartBlurHandler
+} = useFocusMode()
+const {
+	focusMode: thumbEndFocusMode,
+	focusHandler: thumbEndFocusHandler,
+	blurHandler: thumbEndBlurHandler
+} = useFocusMode()
 
 const numberDisplayMousedownHandler = () => {
 	setTimeout(() => {
@@ -638,6 +588,16 @@ onMounted(() => {
 	})
 })
 
+watch([range, disabledComputed, readonlyComputed], (v, o) => {
+	const [range, disabled, readonly] = v
+	const [rangeOld] = o
+	if (disabled || readonly || range !== rangeOld) {
+		thumbFocusMode.value = false
+		thumbStartFocusMode.value = false
+		thumbEndFocusMode.value = false
+	}
+})
+
 watch(
 	[
 		disabledComputed,
@@ -651,7 +611,10 @@ watch(
 		markPoints,
 		modelValue,
 		() => props.direction,
-		() => props.reverse
+		() => props.reverse,
+		thumbFocusMode,
+		thumbStartFocusMode,
+		thumbEndFocusMode
 	],
 	() => {
 		nextTick(() => {
@@ -713,15 +676,17 @@ const drawPixel = () => {
 		drawThumb(
 			thumbRef,
 			thumbCanvasRef,
+			thumbFocusMode.value,
 			thumbStartRef,
 			thumbStartCanvasRef,
+			thumbStartFocusMode.value,
 			thumbEndRef,
 			thumbEndCanvasRef,
+			thumbEndFocusMode.value,
 			props.range,
 			rad,
 			pixelSize,
-			thumbColor,
-			borderColor
+			thumbColor
 		)
 	}