+++ packages/web-vue/lib/switch/index.vue
@@ -8,10 +8,9 @@
 			[`px-switch__active`]: progress > MID_PROGRESS,
 			[`px-switch__inactive`]: progress <= MID_PROGRESS
 		}"
-		ref="switchRef"
-		@mousedown="mousedownHandler"
-		@focusin="focusinHandler"
-		@focusout="focusoutHandler"
+		@mousedown="wrapperMousedownHandler"
+		@focusin="focusHandler"
+		@focusout="blurHandler"
 	>
 		<div
 			class="px-switch-prefix-wrapper"
@@ -102,7 +101,7 @@ import { inBrowser } from '../share/util/env'
 import { useTransitionEnd } from '../share/hook/use-transition-end'
 import { INTERVAL } from '../share/const/style'
 import { usePolling } from '../share/hook/use-polling'
-import { useCancelableDelay } from '../share/hook/use-cancelable-delay'
+import { useFocusMode } from '../share/hook/use-focus-mode'
 import { debounce, isNullish } from 'parsnip-kit'
 
 const MID_PROGRESS = 0.5
@@ -128,7 +127,7 @@ const darkMode = useDarkMode()
 
 const canvasRef = shallowRef<HTMLCanvasElement | null>(null)
 const buttonCanvasRef = shallowRef<HTMLCanvasElement | null>(null)
-const switchButtonRef = shallowRef<HTMLLabelElement | null>(null)
+const switchButtonRef = shallowRef<HTMLDivElement | null>(null)
 const canvasWrapperRef = shallowRef<HTMLDivElement | null>(null)
 
 const [modelValue, updateModelValue] = useControlledMode('modelValue', props, emits, {
@@ -176,34 +175,22 @@ const changeHandler = (e: Event) => {
 	formItemProvide?.changeHandler()
 }
 
-const [wait, cancel] = useCancelableDelay()
-const focusState = ref(false)
 const checkboxRef = shallowRef<HTMLInputElement | null>(null)
 
-const mousedownHandler = () => {
-	setTimeout(() => {
-		checkboxRef.value?.focus()
-	}, 0)
-}
-
-const focusoutHandler = async (e: FocusEvent) => {
-	const next = await wait()
-	if (!next) {
-		return
-	}
-
-	emits('blur', e)
-	formItemProvide?.blurHandler()
-}
-
-const focusinHandler = (e: FocusEvent) => {
-	cancel()
-	const currentFocus = focusState.value
-	focusState.value = true
-	if (!currentFocus) {
-		emits('focus', e)
-	}
-}
+const { focusHandler, blurHandler, wrapperMousedownHandler } = useFocusMode(
+	{
+		onFocus: (e, isFirstFocus) => {
+			if (isFirstFocus) {
+				emits('focus', e)
+			}
+		},
+		onBlur: (e) => {
+			emits('blur', e)
+			formItemProvide?.blurHandler()
+		}
+	},
+	checkboxRef
+)
 
 onMounted(() => {
 	nextTick(() => {