+++ packages/web-vue/lib/textarea/test/focus.test.ts
@@ -1,0 +1,70 @@
+import { mount } from '@vue/test-utils'
+import Textarea from '../index.vue'
+import { vi, describe, it, expect, afterEach, beforeEach } from 'vitest'
+import { createMocks } from '../../share/util/test'
+import { wait } from 'parsnip-kit'
+
+const mountComponent = (props = {}, slots = {}) => {
+	return mount(Textarea, {
+		props: {
+			placeholder: 'Please enter',
+			...props
+		},
+		slots,
+		attachTo: 'body'
+	})
+}
+
+describe('Textarea focus/blur behavior', () => {
+	const { pre, post } = createMocks()
+
+	afterEach(() => {
+		post()
+	})
+	beforeEach(() => {
+		pre()
+	})
+
+	it('wrapper onFocus/onBlur callbacks are called on focus/blur', async () => {
+		const onFocus = vi.fn()
+		const onBlur = vi.fn()
+		const wrapper = mountComponent({ onFocus, onBlur })
+		const input = wrapper.find('textarea.px-textarea-inner')
+
+		await input.trigger('focus')
+
+		expect(onFocus).toHaveBeenCalledTimes(1)
+
+		await input.trigger('blur')
+		await wait(200)
+		expect(onBlur).toHaveBeenCalledTimes(1)
+	})
+
+	it('mousedown on wrapper trigger focus', async () => {
+		const onFocus = vi.fn()
+		const wrapper = mountComponent({ onFocus })
+
+		const el = wrapper.find('.px-textarea')
+
+		await el.trigger('mousedown')
+		await wait(20)
+		expect(onFocus).toBeCalledTimes(1)
+	})
+
+	it('mousedown twice on wrapper trigger only one focus', async () => {
+		const onFocus = vi.fn()
+		const onBlur = vi.fn()
+		const wrapper = mountComponent({ onFocus, onBlur })
+
+		const el = wrapper.find('.px-textarea')
+
+		await el.trigger('mousedown')
+		await wait(20)
+		expect(onFocus).toBeCalledTimes(1)
+
+		await el.trigger('mousedown')
+		await wait(20)
+		expect(onFocus).toBeCalledTimes(1)
+		expect(onBlur).toBeCalledTimes(0)
+	})
+})